tkinter_app/requirements.txt
openai>=1.3.5
python-dotenv>=1.0.0